Refactor graph extras into separate directory.
